refactor: model, lib 함수들 파일 분리, 배럴 파일 적용
