+++ src/entities/comment/api/index.ts
@@ -1,0 +1,4 @@
+export * from "./likeComment.ts"
+export * from "./updateComment.ts"
+export * from "./addComment.ts"
+export * from "./deleteComment.ts"
