+++ src/entities/comment/types/index.ts
