+++ src/entities/post/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./SelectPostFilter.tsx"
+export * from "./Tag.tsx"
