+++ src/entities/user/api/index.ts
@@ -1,7 +1,1 @@
-import { UserResponse } from "../types"
-
-export const fetchUsers = async () => {
-  const response = await fetch("/api/users?limit=0&select=username,image")
-  const data: UserResponse = await response.json()
-  return data.users
-}
+export * from "./fetchUsers.ts"
