+++ src/features/comment/ui/index.ts
@@ -1,0 +1,4 @@
+export * from "./Comment.tsx"
+export * from "./CommentAdd.tsx"
+export * from "./Comment.tsx"
+export * from "./Comment.tsx"
