+++ src/features/post/lib/index.ts
@@ -1,4 +1,4 @@
-import { fetchPosts, searchPost, searchPostByTag } from "@/entities/post/api"
+import { fetchPosts, fetchPostBySearch, fetchPostByTag } from "@/entities/post/api"
 import { fetchUsers } from "@/entities/user/api"
 import { PostResponse, PostWithUser, SearchParams } from "@entities/post/types"
 import { User } from "@entities/user/types"
@@ -18,11 +18,11 @@ export const getPostsWithUsers = async (params: SearchParams) => {
 }
 
 export const getPostsWithUserByTag = async (tag: string) => {
-  const [postsData, usersData] = await Promise.all([searchPostByTag(tag), fetchUsers()])
+  const [postsData, usersData] = await Promise.all([fetchPostByTag(tag), fetchUsers()])
   return combinePostsWithUser(postsData, usersData)
 }
 
 export const getPostsWithUserBySearch = async (searchQuery: string) => {
-  const [postsData, usersData] = await Promise.all([searchPost(searchQuery), fetchUsers()])
+  const [postsData, usersData] = await Promise.all([fetchPostBySearch(searchQuery), fetchUsers()])
   return combinePostsWithUser(postsData, usersData)
 }
