+++ src/features/post/ui/PostAdd.tsx
@@ -10,6 +10,15 @@ export function PostAdd() {
   const { mutate: addPost } = useAddPostMutation()
   const { closeModal } = useModalStore()
 
+  const handleChangeTitle = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setNewPost({ ...newPost, title: e.target.value })
+  }
+  const handleChangeBody = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    setNewPost({ ...newPost, body: e.target.value })
+  }
+  const handleChangeId = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setNewPost({ ...newPost, userId: Number(e.target.value) })
+  }
   const handleAddPost = () => {
     if (!newPost) return
     addPost(newPost)
@@ -22,23 +31,9 @@ export function PostAdd() {
         <DialogTitle>새 게시물 추가</DialogTitle>
       </DialogHeader>
       <div className="space-y-4">
-        <Input
-          placeholder="제목"
-          value={newPost.title}
-          onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
-        />
-        <Textarea
-          rows={30}
-          placeholder="내용"
-          value={newPost.body}
-          onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
-        />
-        <Input
-          type="number"
-          placeholder="사용자 ID"
-          value={newPost.userId}
-          onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
-        />
+        <Input placeholder="제목" value={newPost.title} onChange={handleChangeTitle} />
+        <Textarea rows={30} placeholder="내용" value={newPost.body} onChange={handleChangeBody} />
+        <Input type="number" placeholder="사용자 ID" value={newPost.userId} onChange={handleChangeId} />
         <Button onClick={handleAddPost}>게시물 추가</Button>
       </div>
     </DialogContent>
