+++ src/features/user/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserModal.tsx"
