+++ src/shared/ui/index.ts
@@ -4,3 +4,5 @@ export * from "./dialog"
 export * from "./inputs"
 export * from "./table"
 export * from "./thypograph"
+export * from "./loading"
+export * from "./modal"
