+++ src/shared/ui/loading/Loading.tsx
@@ -1,0 +1,3 @@
+export function Loading() {
+  return <div className="flex justify-center p-4">로딩 중...</div>
+}
