+++ src/shared/ui/loading/index.ts
@@ -1,0 +1,1 @@
+export * from "./Loading.tsx"
