+++ src/widgets/ui/Footer.tsx
@@ -7,5 +7,3 @@ export function Footer() {
     </footer>
   )
 }
-
-export default Footer
