+++ src/widgets/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./Header.tsx"
 export * from "./Footer.tsx"
+export * from "./Pagination.tsx"
